Upload code for step1 MapReduce
This is the final version of step1 MapReduce code. 

It's revised by the fact that we should not to keep all the keys and values in one place, because all the items from the step1 mapper have been already sorted by its key. All the items with same key will be sent to the same reducer. In doing so, we can avoid running out of memory when executing step1 MapReduce. 

And also all the outliers and invalid pages will be filtered out in this version code.