Upload code for step2 MapReduce
This is the final version of step2 MapReduce.

Please note that when executing the step2 MapReduce on AWS EMR, we may need some arguments like: 
-D stream.num.map.output.key.fields=2
-D map.output.key.field.separator=}
-D mapreduce.partition.keypartitioner.options=-k1,1
-partitioner org.apache.hadoop.mapred.lib.KeyFieldBasedPartitioner

In doing so, we can make sure all the step2 map result with same page name will be sent to the same reducer of step2, so we can guarantee that all the unpopular pages with page views less than 10 will be filtered out correctly.  

And again, this revised version code leverage the fact that intermediate data from mapper are sorted, so we don't need to store all the keys and values from mapper in one place, and it saves a lot of memory space.